chore: MockSession extension에 위치
